Rename main .py file, change readme, fix readme typo
