+++ README.md
@@ -1,5 +1,5 @@
-# ytdl-downloader
-A small terminal based gui program that downloads youtube videos in mp4 or mp3 formats. It uses <a href="https://youtube-dl.org/" target="_blank">youtube-dl</a> and <a href="https://ffmpeg.org/" target="_blank">ffmpeg tools</a>. The .exe version of the program was compiled using <a href="https://pyinstaller.org/en/stable/" target="_blank">PyInstaller</a>.
+# Youtube-dl Terminal Gui
+Previously called *ytdl-downloader*, youtube-dl terminal gui (or **ytdl-tgui** for short) is a small terminal based gui program that downloads youtube videos in mp4 or mp3 formats. It uses <a href="https://youtube-dl.org/" target="_blank">youtube-dl</a> and <a href="https://ffmpeg.org/" target="_blank">ffmpeg tools</a>. The .exe version of the program was compiled using <a href="https://pyinstaller.org/en/stable/" target="_blank">PyInstaller</a>.
 
 For more information about any errors you may encounter when using the program, I encourage you go look up the youtube-dl documentation <a href="https://github.com/ytdl-org/youtube-dl/" target="_blank">here</a> since this program is just superimposed over youtube-dl.
 
@@ -8,17 +8,17 @@ For more information about any errors you may encounter when using the program, 
 
 <br>
 
-## Why should I use this?
+## Why should I use this? 💻
 To be honest, you don't really need to. The youtube-dl command tool can pretty much achieve everything this program can, however, this program is not meant to replace youtube-dl.
 
-So, *"why should I use this?"* you may ask. Well, if you know youtube-dl commands by heart, you don't really need this program, but for those who are not familiar with the youtube-dl command tool, and for those who wake up tired after an 8 hour coding session and want to download something quick, this program is meant for you. 
+So, *"why should I use this?"* you may ask. Well, if you know how to use youtube-dl by heart, you don't really need this program, but for those who are not familiar with the youtube-dl command tool, and for those who wake up tired after an 8 hour coding session and want to download something quick, this program is meant for you. 
 
-Yt-dl downloader simplifies each step of the download for a more user-friendly experience so you don't screw up badly and end up downloading *never gonna give you up* in 144p who knows where.
+**Youtube-dl terminal gui** simplifies each step of the download for a more user-friendly experience so you don't screw up badly and end up downloading *never gonna give you up* in 144p who knows where in your machine.
 
-In the future, I plan on adding an option to allow you to input your own youtube-dl commands inside the program; a feature I believe some of you will be glad it exist. Although, you know, if you ever feel limited by the constraints of thi program, you can always hope back to using youtube-dl in the terminal.
+In the future, I plan on adding an option to allow you to input your own youtube-dl commands inside the program; a feature I believe some of you will be glad it exist. Although, you know, if you ever feel limited by the constraints of this program, you can always hope back to using youtube-dl in the terminal.
 
-## How does it work
-Yt-dl downloader is simply a kind of gui wrapper of the youtube-dl command tool. A good way to explain how it works would be like this:
+## How does it work ⚙️
+**Youtube-dl terminal gui** is simply a kind of gui wrapper of the youtube-dl command tool. A good way to explain how it works would be like this:
 
 <br>
 <div align="center">
@@ -26,7 +26,7 @@ Yt-dl downloader is simply a kind of gui wrapper of the youtube-dl command tool.
 </div>
 
 <br>
-So basically, any command that works in youtube-dl works on this program too. If you check the .py file of yt-dl downloader, you may find that the program works by simply running youtube-dl commands in the background. As I said, this program is not meant to replace youtube-dl, is simply an extension of it.
+So basically, any command that works in youtube-dl works on this program too. If you check the .py file ytdl-terminal-gui, you may find that the program works by simply running youtube-dl commands in the background. As I said, this program is not meant to replace youtube-dl, is simply an extension of it.
 
 <br>
 
@@ -37,7 +37,7 @@ There are two ways to open the program:
 
 ```bash
 # Opens file using your python version
-python -m location/of/.py/file
+python3 -m location/of/.py/file
 ```
 <br>
 
@@ -53,7 +53,7 @@ open location/of/.exe./file
 
 
 ## Dependencies 📊
-You need to have youtube-dl and ffmpeg tools installed. Otherwise, both the .py and .exe versions of the program won't work. You can installed them with pip using the following commands:
+You need to have youtube-dl and ffmpeg tools installed. Otherwise, both the .py and .exe versions of the program won't work. You can install them with pip using the following commands:
 
 ```bash
 pip install youtube_dl
@@ -63,18 +63,20 @@ pip install ffmpeg
 ```
 <br>
 
+Also, **youtube-dl terminal gui** is made with python 3, so make sure to use the python version from v3 to upwards. 
+
 Don't have pip? You can download it running this command:
 <br>
 <h3>Linux & MacOs</h3>
 
 ```bash
-python -m ensurepip --upgrade
+python3 -m ensurepip
 ```
 
 <h3>Windows</h3>
 
 ```bash
-py -m ensurepip --upgrade
+py -m ensurepip
 ```
 <br>
 
@@ -87,7 +89,7 @@ For some reason I don't understand, I cannot get PyInstaller to compile multiple
 
 I don't know what to do, I'm tired, and for some reason PyInstaller just refuses to work. So, for now, I'm leaving it like that until I find a way how to compile a .py file that imports from other .py files.
 
-If by chance you know how to do it, feel free to make a pull request, or open a issue, or to reach me out, either way I'll be happy for any kind of contribution regarding this matter. Although, if you want to contribute in any way or form, feel free to do so by making a pull request or opening a issue, any kind of help will be gladly recieved 😁.
+If by chance you know how to do it, feel free to make a pull request, or open a issue, or create a discussion, or just reach me out, either way, I'll be happy for any kind of help regarding this matter. Although, if you want to contribute in any way or form, feel free to do so, any kind of help will be gladly recieved 😁.
 
 <br>
